custom font added; meta tags added;favicon update
