+++ .github/CODE_OF_CONDUCT.md
@@ -143,10 +143,9 @@ version 2.0, available at
 Community Impact Guidelines were inspired by [Mozilla's code of conduct
 enforcement ladder](https://github.com/mozilla/diversity).
 
-[homepage]: https://www.contributor-covenant.org
-
 For answers to common questions about this code of conduct, see the FAQ at
 <https://www.contributor-covenant.org/faq>. Translations are available at
 <https://www.contributor-covenant.org/translations>.
 
 [contact]: mailto:xkeshav@gmail.com
+[homepage]: https://www.contributor-covenant.org
