+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this repository will be documented in this file.
 
-## [0.0.1.0]
+## [0.0.1]
 
 - Initial release.
+
+## [1.0.0]
+
+- Custom Font added
+- Meta tags added
