+++ src/assets/styles/alphabet.css
@@ -3,6 +3,7 @@
   grid-gap: 4px;
   container-type: inline-size;
   container-name: layout;
+  font-family: 'atkinson';
 }
 
 .list {
