+++ src/assets/styles/header.css
@@ -1,4 +1,3 @@
-@import url('./global.css');
 
 .header {
   background-color: #15091d;
