+++ src/pages/index.astro
@@ -7,8 +7,8 @@ import Navigation from "@/layouts/Navigation";
 import { pages } from "@/utils/constant";
 ---
 
-<Layout title="abcdkbd.com: kidz">
-  <Header title="Kid lab" hasBack={false}>
+<Layout title="parixan: abcd">
+  <Header title="Parixan" hasBack={false}>
     <div class="switch">dark</div>
   </Header>
 	<main>
