+++ src/pages/varnmala.astro
@@ -3,6 +3,7 @@ import Alphabet from "@/components/Alphabet";
 import Header from "@/components/Header";
 import Layout from "@/layouts/Layout";
 import { varnmala_english, varnmala_hindi } from '@/utils/common';
+
 ---
 
 <Layout title="varnmala" >
