+++ src/types/base.ts
@@ -1,0 +1,14 @@
+export type SiteMeta = {
+  title: string;
+  ogImage?: string;
+};
+
+export type SiteConfig = {
+  website: string;
+  author: string;
+  title: string;
+  description: string;
+  language: string;
+  ogLocale: string;
+  date: any;
+};
