+++ src/types/index.ts
@@ -1,10 +1,3 @@
-export type SiteConfig = {
-  website: string;
-  author: string;
-  title: string;
-  description: string;
-  language: string;
-  date: any;
-};
-
 export type { LetterEntity } from "./alphabet";
+
+export type { SiteConfig, SiteMeta } from "./base";
